+++ src/features/ChangeButtonSize/hooks/useButtonSize.ts
@@ -1,0 +1,16 @@
+import { buttonSizeState } from '@/stores/buttonSize';
+import { Size } from '@/types/button';
+import { useSetRecoilState } from 'recoil';
+
+function useButtonSize() {
+  const setButtonSize = useSetRecoilState(buttonSizeState);
+
+  const handleCheckBoxChange = (size: Size) => {
+    setButtonSize(size);
+  };
+  return {
+    handleCheckBoxChange,
+  };
+}
+
+export default useButtonSize;
